Loads don't alias with any other loads now
This means that equivalent loads have identical preds and can be removed
by CSE.
